add TileSetCodeSupport
